Continue CORE-382 - Signals Loading
How it currently works
User loads the landing page
The frontend fetches the latest asset data from the ocean APIs
The assets are fetched from the ocean APIs and returned
Because we have added the signals provider to the home page we have our signals useEffect listening for when the assets have been loaded (Only for the highest liquidity query for testing purposes )
Once the assets are loaded we can use their IDs to fetch the signals for each asset available
We join all available assetIds(DIDs) with the origin/source url for the signal hook to fetch the signal information for the assets
We fetch signals from a (mock) API and with a static  origin/source url
We wait for the signals to load and then log them in the console to review the data available. (Ideally the signals would be ready to be displayed on the signal components)

How it should work
User sets their signal settings in the settings page
Which signals they would like to see, which pages the signals should be shown on e.g only on list view or detail view
The API automatically saves the signal settings to the localStorage and within user preferences
User loads the relevant page
The users signal settings are loaded
The frontend fetches the latest asset data from the ocean APIs
The assets are fetched from the ocean APIs and returned
Because we have added the signals provider to the relevant page we have useEffect hooks listening for when the assets have been loaded
Once the assets are loaded we can use their IDs and any other signal url parameters to fetch the signals for each asset available
Because signal settings are stored locally we know what signals to fetch for on this the current loaded page
We join all available assetIds(DIDs) , publisherIds, userAddresses with various origin/source urls from the user signal settings and use the signal hook in the context provider to fetch the signal information for the assets
We fetch signals from multiple (mock) APIs based on the user's signal preferences
We wait for the signals to load and then return them to be displayed by the signal components

TODO

Refactor hooks for signals fetch multiple SWR calls based on settings parameters
How do we best call multiple useSWR calls to fetch an arbitrary number of signal requests once all required params (assetIds, publisherIds etc) have been loaded
How do we best restructure the useSignals hook for loading the signals from the settings
Store signal settings preferences in user preferences
Fetch asset signals based on user signal settings
Confirm the requirements for  creating new custom signals i.e
Are we going to need a signal type?
Is there a signal icon to be stored or provided by the signal API or user?
Are we expecting any number of paths from custom signal URLs with multiple assetIds, publisherIds, userAccounts?